Move TypeScript code to typescript.ts (no code changes)
